+++ src/main/java/com/finco/finco/infrastructure/account/dto/AccountTransactionData.java
@@ -1,7 +1,9 @@
 package com.finco.finco.infrastructure.account.dto;
 
+import java.math.BigDecimal;
+
 import com.finco.finco.usecase.account.dto.IAccountTransactionData;
 
-public record AccountTransactionData(Long amount) implements IAccountTransactionData{
+public record AccountTransactionData(BigDecimal amount) implements IAccountTransactionData{
 
 }
