+++ src/main/java/com/finco/finco/usecase/account/dto/IAccountPublicData.java
@@ -1,5 +1,6 @@
 package com.finco.finco.usecase.account.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.finco.finco.entity.account.model.AccountType;
@@ -11,7 +12,7 @@ public interface IAccountPublicData {
     Long userId();
     String name();
     AccountType type();
-    Long balance();
+    BigDecimal balance();
     CurrencyEnum currency();
     LocalDateTime creationDate();
     String description();
