+++ src/main/java/com/finco/finco/usecase/account/dto/IAccountRegistrationData.java
@@ -1,5 +1,7 @@
 package com.finco.finco.usecase.account.dto;
 
+import java.math.BigDecimal;
+
 import com.finco.finco.entity.account.model.AccountType;
 import com.finco.finco.entity.account.model.CurrencyEnum;
 
@@ -8,7 +10,7 @@ public interface IAccountRegistrationData {
     Long userId();
     String name();
     AccountType type();
-    Long balance();
+    BigDecimal balance();
     CurrencyEnum currency();
     String description();
